Add requirement and environ setting in root directory
Before this change the django apps(`blog`, `blog_manager`, `loperuv`)
are in the root directory. It's not easy to use when we want to
build the site on a new machine.
So in this change, we add `requirement.txt`, `env.sh` and so forth
environ setting files in the root directory, and make the django
project directory in the root directory, like this:

-README.md // will be added later
-env.sh
-requirement.txt
-yishan_toys/ // which is the new name of the site
----loperuv/
----blog/
----blog_manager/
----manage.py // generated by django-admin

So we can build the project.